Fix #68441: wrong ZPP for svn_diff()
If a `zend_long` parameter is nullable, it needs an additional
`zend_bool` argument which indicates whether the passed argument
actually was `NULL`; otherwise the parameter memory will be corrupted.

Since the default value of these parameters is `SVN_REVISION_HEAD`
anyway, there is no need for the parameters to be nullable at all.
This also matches the "proto" and the docs.  This is also not a real
BC break, since passing `NULL` to these parameters likely crashes
anyway.

The patch has been provided by the bug reporter.